requirements.txt
pygame
farm-ng-amiga>=0.0.9